introduce 'iframe' mode
Until now it was necessary for the player to work to be embedded via a
<script> tag. Now it is also possible to do the embed with a <iframe>
tag. This comes with the constraint that the player size can't change
dynamically. So either one of the extension panels needs to be opened on
start via configuration, or no panels will be allowed to open at all.